requirements.txt
asgiref==3.5.2
attrs==22.2.0
autobahn==22.12.1
Automat==22.10.0
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.1
colorama==0.4.6
commonmark==0.9.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
daphne==3.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.2.6
Django==4.1.5
django-allauth==0.52.0
django-dotenv==1.4.2
django-eventstream==4.5.0
django-grip==3.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
gripcontrol==4.1.0
gunicorn==20.1.0
httpie==3.2.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.4
mysqlclient==2.1.1
oauthlib==3.2.2
packaging==22.0
pubcontrol==3.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.16.0
Pygments==2.14.0
PyJWT==2.6.0
pyOpenSSL==22.1.0
PySocks==1.7.1
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rich==13.0.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
Werkzeug==2.0.3
zope.interface==5.5.2
